speech/chatterbox-tts-api/pyproject.toml
[project]
name = "chatterbox-tts-api"
version = "2.1.0"
description = "REST API for Chatterbox TTS with OpenAI compatibility"
authors = [{ name = "Travis Van Nimwegen", email = "tts@travis2.com" }]
readme = "README.md"
license = { text = "AGPLv3" }
requires-python = ">=3.11, <3.12"
dependencies = [
  # "chatterbox-tts @ git+https://github.com/resemble-ai/chatterbox.git",
  "chatterbox-tts @ git+https://github.com/travisvn/chatterbox-multilingual.git@exp",
  "resemble-perth",
  "torch>=2.0.0,<2.7.0",
  "torchaudio>=2.0.0,<2.7.0",
  "fastapi>=0.104.0",
  "uvicorn[standard]>=0.24.0",
  "pydantic>=2.0.0",
  "python-multipart>=0.0.6",
  "python-dotenv>=1.0.0",
  "psutil>=5.9.0",
  "requests>=2.28.0",
  "sse-starlette>=3.0.2",
  "pydub>=0.25.1",
]

[project.urls]
Homepage = "https://github.com/travisvn/chatterbox-tts-api"
Repository = "https://github.com/travisvn/chatterbox-tts-api"

[project.optional-dependencies]
dev = [
  "requests>=2.28.0", # for testing
]
test = [
  "pytest>=7.0.0",
  "pytest-asyncio>=0.21.0",
  "pytest-cov>=4.0.0",
  "pytest-html>=3.0.0",
  "pytest-xdist>=3.0.0",
  "requests>=2.28.0",
  "httpx>=0.24.0",
]

[tool.uv]
dev-dependencies = ["requests>=2.28.0"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = [
  "-ra",
  "--strict-markers",
  "--strict-config",
  "--cov=app",
  "--cov-report=term-missing",
  "--cov-report=html",
  "--cov-report=xml",
]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
  "slow: marks tests as slow (deselect with '-m \"not slow\"')",
  "integration: marks tests as integration tests",
  "unit: marks tests as unit tests",
  "api: marks tests as API tests",
  "memory: marks tests as memory management tests",
  "voice: marks tests as voice-related tests",
  "regression: marks tests as regression tests",
]
